+++ services/api-service/src/apps/debttrack/debt/debt.repository.ts
@@ -1,15 +1,17 @@
 import { Injectable } from "@nestjs/common"
-import { InjectModel } from "@nestjs/mongoose"
 import { Debt } from "./schemas/debt.schema"
 import { AppsDbConnectionMap } from "@/shared/constants/db-connection.map"
-import { Model } from "mongoose"
-import { EntityRepository } from "@/shared/entity/entity.repository"
+import {
+  EntityRepository,
+  EntityModel,
+  InjectEntityModel,
+} from "@/shared/entity/entity.repository"
 
 @Injectable()
 export class DebtRepository extends EntityRepository<Debt> {
   constructor(
-    @InjectModel(Debt.name, AppsDbConnectionMap.DebtTrack)
-    private debtModel: Model<Debt>
+    @InjectEntityModel(Debt.name, AppsDbConnectionMap.DebtTrack)
+    private debtModel: EntityModel<Debt>
   ) {
     super(debtModel)
   }