+++ services/api-service/src/apps/expensetrack/expense/expense.repository.ts
@@ -1,15 +1,17 @@
 import { Injectable } from "@nestjs/common"
-import { InjectModel } from "@nestjs/mongoose"
 import { Expense } from "./schemas/expense.schema"
 import { AppsDbConnectionMap } from "@/shared/constants/db-connection.map"
-import { Model } from "mongoose"
-import { EntityRepository } from "@/shared/entity/entity.repository"
+import {
+  EntityModel,
+  EntityRepository,
+  InjectEntityModel,
+} from "@/shared/entity/entity.repository"
 
 @Injectable()
 export class ExpenseRepository extends EntityRepository<Expense> {
   constructor(
-    @InjectModel(Expense.name, AppsDbConnectionMap.ExpenseTrack)
-    private expenseModel: Model<Expense>
+    @InjectEntityModel(Expense.name, AppsDbConnectionMap.ExpenseTrack)
+    private expenseModel: EntityModel<Expense>
   ) {
     super(expenseModel)
   }