+++ services/api-service/src/apps/wealthgoal/goal/goal.repository.ts
@@ -1,16 +1,18 @@
 import { Injectable } from "@nestjs/common"
-import { InjectModel } from "@nestjs/mongoose"
 import { Goal } from "./schemas/goal.schema"
 import { AppsDbConnectionMap } from "@/shared/constants/db-connection.map"
-import { Model } from "mongoose"
-import { EntityRepository } from "@/shared/entity/entity.repository"
+import {
+  EntityModel,
+  EntityRepository,
+  InjectEntityModel,
+} from "@/shared/entity/entity.repository"
 import objectId from "@/shared/utils/convert-objectid"
 
 @Injectable()
 export class GoalRepository extends EntityRepository<Goal> {
   constructor(
-    @InjectModel(Goal.name, AppsDbConnectionMap.WealthGoal)
-    private goalModel: Model<Goal>
+    @InjectEntityModel(Goal.name, AppsDbConnectionMap.WealthGoal)
+    private goalModel: EntityModel<Goal>
   ) {
     super(goalModel)
   }