+++ services/api-service/src/platform/analytics/analytics.repository.ts
@@ -1,15 +1,17 @@
 import { Injectable } from "@nestjs/common"
-import { InjectModel } from "@nestjs/mongoose"
 import { Analytics } from "./schemas/analytics.schema"
 import { GeneralDbConnectionMap } from "@/shared/constants/db-connection.map"
-import { Model } from "mongoose"
-import { EntityRepository } from "@/shared/entity/entity.repository"
+import {
+  EntityModel,
+  EntityRepository,
+  InjectEntityModel,
+} from "@/shared/entity/entity.repository"
 
 @Injectable()
 export class AnalyticsRepository extends EntityRepository<Analytics> {
   constructor(
-    @InjectModel(Analytics.name, GeneralDbConnectionMap.Platform)
-    private analyticsModel: Model<Analytics>
+    @InjectEntityModel(Analytics.name, GeneralDbConnectionMap.Platform)
+    private analyticsModel: EntityModel<Analytics>
   ) {
     super(analyticsModel)
   }