+++ services/api-service/src/platform/subscription/repositories/subscription.repository.ts
@@ -1,15 +1,17 @@
 import { Injectable } from "@nestjs/common"
-import { InjectModel } from "@nestjs/mongoose"
 import { Subscription } from "../schemas/subscription.schema"
 import { GeneralDbConnectionMap } from "@/shared/constants/db-connection.map"
-import { Model } from "mongoose"
-import { EntityRepository } from "@/shared/entity/entity.repository"
+import {
+  EntityModel,
+  EntityRepository,
+  InjectEntityModel,
+} from "@/shared/entity/entity.repository"
 
 @Injectable()
 export class SubscriptionRepository extends EntityRepository<Subscription> {
   constructor(
-    @InjectModel(Subscription.name, GeneralDbConnectionMap.Platform)
-    private subscriptionModel: Model<Subscription>
+    @InjectEntityModel(Subscription.name, GeneralDbConnectionMap.Platform)
+    private subscriptionModel: EntityModel<Subscription>
   ) {
     super(subscriptionModel)
   }